requirements.txt
flask==3.0.3
gunicorn==22.0.0
python-dotenv==1.0.1
openai==1.40.0
chromadb==0.5.5
sentence-transformers==3.0.1
langchain==0.2.12
langchain-openai==0.1.20
langchain-community==0.2.11
langchain-huggingface==0.0.3
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
joblib==1.4.2
xgboost==2.1.1
pypdf==pypdf-6.10.2